refactor(auth): extract setup seam
Move first-run setup + credential management to a dedicated module:

- USERNAME_MIN_LENGTH, USERNAME_MAX_LENGTH, _USERNAME_PATTERN
- _setup_complete (lazy monotonic cache)
- is_setup_complete, set_password
- rotate_session_secret (composes session.reset_serializer)
- normalize_username, validate_username
- set_username, get_username
- check_password, check_credentials
- reset_auth_caches (composes session.reset_serializer and
  rate_limit.reset_login_attempts; only cross-seam orchestrator
  that knows every seam's reset helper)

auth/__init__.py re-exports every name; __getattr__ now routes
``_setup_complete`` and ``_USERNAME_PATTERN`` pass-throughs so
``houndarr.auth._setup_complete`` always reads the live value
owned by setup.py.  Drops two now-unused imports from the
package root (get_setting, set_setting, compare_digest stays
because csrf + proxy_auth sections still need it).
